CDC-15 Refactor: react-query & sort libs
