+++ apps/native/libs/react-query/item.ts
@@ -1,0 +1,3 @@
+export const itemQueryKeys = {
+  byId: (id: number, userId?: string) => ['items', 'byId', { id, userId }],
+};
