+++ apps/native/libs/sort/index.ts
@@ -1,4 +1,5 @@
 export * from './discount';
+export * from './my-comment';
 export * from './ranking';
 export * from './search';
 export * from './types';
